refactor(native): group runtime parallel and interlock modules
