+++ src/native/engine/CMakeLists.txt
@@ -88,11 +88,11 @@ set(ENGINE_RUNTIME_SOURCES
   runtime/merchant/merchant_trade_domain.c
   runtime/merchant/merchant_service.c
   runtime/interactive_objects.c
-  runtime/interlock.c
+  runtime/parallel/interlock.c
   runtime/input/move_target_service.c
   runtime/input/move_target_domain.c
-  runtime/parallel_contract.c
-  runtime/parallel_model.c
+  runtime/parallel/parallel_contract.c
+  runtime/parallel/parallel_model.c
   runtime/physics_abi.c
   runtime/player/player_api.c
   runtime/player/player_merchant_adapter.c
