+++ tests/native/runtime_parallel_amdahl_test.c
@@ -1,4 +1,4 @@
-#include "runtime/parallel_model.h"
+#include "runtime/parallel/parallel_model.h"
 
 #include <math.h>
 #include <stdio.h>
